fix(tx-dma-trans-buffer): derive channel index from metadata
Previously, the channel index has been derived from the PCIe address
which is okay if the channel number is encoded in the address itself but
does not allow for custom channel indexing. This has been changed within
this commit in order to provide for custom indexing by a channel index
that gets transported with the SOF of incoming data.