fix(top-level sim): rename signals in PCIe AXI to use the right signal in simulator bus driver
The bus drivers use the same naming/prefix pattern for all signals on the bus.
However, one signal connected directly to the PCIe IP core was named differently.
This caused the bus driver to drive a signal that had already been driven due to a VHDL assignment.

Some simulators have optimised this assignment, so it works anyway, but not all simulations do.