docs(tx-dma-trans-buffer): add implementation note
This implementation note explains the difficulty of implementing the
URAM-based buffer insted of BRAM-based one as is currently implemented.
Although this buffer provides resource saving of RAM blocks by the
factor of 8, it introduces the flaw of appearing write conflicts when
two neighboring DWords need to be written on the same address.